requirements.txt
python-telegram-bot==20.8
openai==0.28
fastapi
uvicorn[standard]
